requirements.txt
dj-database-url==0.4.0
Django==1.9.4
django-braces==1.8.1
django-oauth-toolkit==0.10.0
django-rest-auth==0.7.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.3.3
drf-extensions==0.2.8
gunicorn==19.4.5
oauthlib==1.0.3
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
whitenoise==3.0
